Discord-Rep-Bot/requirements.txt
discord.py
discord-py-interactions